Add includePackageNames parameter support.
The includePackageNames parameter can be set to restrict the packages included in the jdiff descriptor. This is really useful when you only want to describe the public API changes.
The list of packages to include can be provided as a list of packages separated by spaces from both execution configuration and command line.